requirements.txt
alembic==1.10.0
amqp==5.1.1
antiorm==1.2.1
async-timeout==4.0.2
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
db==0.1.1
Deprecated==1.2.13
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Limiter==3.3.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
idna==3.4
init==0.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
limits==3.2.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
Pygments==2.14.0
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
rich==13.3.2
six==1.16.0
SQLAlchemy==2.0.5.post1
terminal==0.4.0
typing_extensions==4.5.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1

Flask-Celery-Helper~=1.1.0
pytest~=7.2.2